@JamesMTucker added data update
I was able to successfully use Snscrape and collected new data. Performed sentiment analysis on the data and confirmed my hypothesis. I wrote a paragraph write up and copied some data. Onto the sheet. It is up to you whether you want to include this in the final grade or not, but I thought it would be good in the repo just in case.